fix: forgot to remove the previous MapLayer & cleaning code
On row 47: + map.removeLayer(this.distortableImage)
Now every time the component is updating it remove the previous leaflet Map Layer. Without this fix if you change the render props it will always add a new Leaflet Layer that increase the memory.

On row 57: there was a '2', it was a typing error I suppose. I deleted it.

On Index.js I removed the corner state because I prefer to pass them to the <ReactDistortableImageOverlayMapLayer /> child component as a props and not as a state.